Reorganise components interfaces and move implementation functions
- ✅ Removed "GoldenLayout Type Safety Infrastructure" banner from components.interfaces.ts
- ✅ Moved GoldenLayout import to top of file with other imports
- ✅ Moved implementation functions from components.interfaces.ts to components.ts:
  - createTypedComponentConfig() - factory function
  - createTypedLayoutItem() - factory function
  - fromGoldenLayoutConfig() - validation/conversion function
  - toGoldenLayoutConfig() - conversion function
  - createTypedDragSource() - wrapper function
  - validateComponentState() and related validation helpers
- ✅ Updated all import statements across 5 files to import functions from components.js instead
- ✅ Kept only interface definitions and type guards in components.interfaces.ts
- ✅ Made validation non-permissive: unknown component names now return false instead of true
- ✅ All TypeScript checks pass, all tests pass (1091 tests), linter compliance achieved

Components.interfaces.ts now contains only true interface definitions while components.ts
contains all implementation functions, providing cleaner separation of concerns.
