+++ static/panes/tree.ts
@@ -30,7 +30,7 @@ import _ from 'underscore';
 import {escapeHTML} from '../../shared/common-utils.js';
 import {assert, unwrap, unwrapString} from '../assert.js';
 import * as BootstrapUtils from '../bootstrap-utils.js';
-import {createTypedDragSource} from '../components.interfaces.js';
+import {createTypedDragSource} from '../components.js';
 import * as Components from '../components.js';
 import {EventHub} from '../event-hub.js';
 import {Hub} from '../hub.js';
